Move generation metadata
