+++ pkg/metadata/generation_metadata.go
@@ -11,7 +11,7 @@
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package ack
+package metadata
 
 import (
 	"crypto/sha1"
